+++ src/core/lang/sv/symbols/program.rs
@@ -1,5 +1,5 @@
 //
-//  Copyright (C) 2022-2025  Chase Ruskin
+//  Copyright (C) 2022-2026  Chase Ruskin
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by