+++ tools/license.py
@@ -4,7 +4,7 @@
 import glob
 
 NEW_HEADER = '''//
-//  Copyright (C) 2022-2025  Chase Ruskin
+//  Copyright (C) 2022-2026  Chase Ruskin
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -23,7 +23,7 @@ NEW_HEADER = '''//
 '''
 
 OLD_HEADER = '''//
-//  Copyright (C) 2022-2024  Chase Ruskin
+//  Copyright (C) 2022-2025  Chase Ruskin
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by