+++ src/commands/env.rs
@@ -121,7 +121,6 @@ impl Env {
                 });
             }
         }
-
         println!("{}", result);
         Ok(())
     }