+++ src/util/filesystem.rs
@@ -240,16 +240,20 @@ pub fn get_exe_path() -> Result<PathBuf, Fault> {
 /// This function is helpful for resolving full paths in plugin arguments,
 /// config.toml includes, and template paths.
 pub fn resolve_rel_path(root: &std::path::PathBuf, s: &str) -> String {
-    let resolved_path = root.join(&s);
-    if std::path::Path::exists(&resolved_path) == true {
-        if PathBuf::from(&s).is_relative() == true {
-            // write out full path
-            PathBuf::standardize(resolved_path).display().to_string()
+    if s.len() == 0 {
+        s.to_string()
+    } else {
+        let resolved_path = root.join(&s);
+        if std::path::Path::exists(&resolved_path) == true {
+            if PathBuf::from(&s).is_relative() == true {
+                // write out full path
+                PathBuf::standardize(resolved_path).display().to_string()
+            } else {
+                s.to_string()
+            }
         } else {
             s.to_string()
         }
-    } else {
-        s.to_string()
     }
 }